feat: hw-health alerts now have level
